requirements.txt
absl-py==0.10.0
altair==4.1.0
astor==0.8.1
attrs==20.3.0
backcall==0.2.0
base58==2.0.0
bleach==3.2.1
blinker==1.4
boto3==1.12.41
botocore==1.15.41
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
dlib==19.18.0
docutils==0.16
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
gast==0.3.3
google-auth==1.17.2
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
matplotlib==3.2.2
Markdown==3.2.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.8
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python-headless==4.4.0.46
opencv-python-headless==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.5
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.3.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2019.3
pywinpty==0.5.7
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
streamlit==0.72.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.9
validators==0.14.3
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
efficientnet==1.1.1